feat: WindowPin
