+++ view/src/components/index.ts
@@ -1,4 +1,5 @@
 import MessageInput from "./MessageInput/index.vue";
 import MessageList from "./MessageList/index.vue";
+import WindowPin from "./WindowPin/index.vue";
 
-export { MessageInput, MessageList };
+export { MessageInput, MessageList, WindowPin };
